xmlreader clarity and copyright
