+++ lemminx-liberty/src/test/java/io/openliberty/XmlReaderTest.java
@@ -10,28 +10,26 @@ import javax.xml.stream.XMLStreamException;
 
 import org.junit.jupiter.api.Test;
 
-import io.openliberty.tools.langserver.lemminx.services.XmlReader;
+import io.openliberty.tools.langserver.lemminx.util.XmlReader;
 
-public class StaxReaderTest {
-    static XmlReader reader = XmlReader.getInstance();
-
+public class XmlReaderTest {
     File resourcesDir = new File("src/test/resources");
 
     @Test
     public void readEmptyXml() throws IOException, XMLStreamException {
         File emptyXml = new File(resourcesDir, "empty_server.xml");
-        assertFalse(reader.hasServerRoot(emptyXml));
+        assertFalse(XmlReader.hasServerRoot(emptyXml));
     }
 
     @Test
     public void readServerXml() throws IOException, XMLStreamException {
         File sampleServerXml = new File(resourcesDir, "sample/server.xml");
-        assertTrue(reader.hasServerRoot(sampleServerXml));
+        assertTrue(XmlReader.hasServerRoot(sampleServerXml));
     }
 
     @Test
     public void readLibertPluginConfigXml() throws IOException, XMLStreamException {
         File lpcXml = new File(resourcesDir, "sample/liberty-plugin-config.xml");
-        assertFalse(reader.hasServerRoot(lpcXml));
+        assertFalse(XmlReader.hasServerRoot(lpcXml));
     }
 }
